add access petani

+++ routes/farmRouter.js
@@ -6,29 +6,38 @@ const fetch = require("node-fetch");
 farmRouter
   .route("/")
   .get(async (req, res, next) => {
-    try {
-      const farms = await Farm.find();
-      res.json(farms);
-    } catch (error) {
-      const err = new Error(error);
-      next(err);
+    if(req.user.accountType === 'Petani'){
+      try {
+        const farms = await Farm.find();
+        res.json(farms);
+      } catch (error) {
+        const err = new Error(error);
+        next(err);
+      }
+    } else{
+      res.status(400).json({ error: "Token is not valid" });
     }
   })
   .post(async (req, res, next) => {
-    try {
-      const farm = await Farm.findOne({ name: req.body.name });
-      if (farm) {
-        const err = new Error("Farm already exists.");
-        err.status = 400;
+    if(req.user.accountType === 'Petani'){
+      try {
+        const farm = await Farm.findOne({ name: req.body.name });
+        if (farm) {
+          const err = new Error("Farm already exists.");
+          err.status = 400;
+          next(err);
+        }
+        const newFarm = new Farm(req.body);
+        await newFarm.save();
+        res.status(200).json({ message: "Farm registered successfully!" });
+      } catch (error) {
+        const err = new Error(error);
         next(err);
       }
-      const newFarm = new Farm(req.body);
-      await newFarm.save();
-      res.status(200).json({ message: "Farm registered successfully!" });
-    } catch (error) {
-      const err = new Error(error);
-      next(err);
+    } else {
+      res.status(400).json({ error: "Token is not valid" });
     }
+  
   })
   .put((req, res, next) => {
     const err = new Error("PUT method is not allowed.");
@@ -44,14 +53,20 @@ farmRouter
 farmRouter
   .route("/:id")
   .get(async (req, res, next) => {
-    try {
-      const farm = await Farm.findById(req.params.id);
-      res.json(farm);
-    } catch (error) {
-      const err = new Error("Farm not found.");
-      err.status = 404;
-      next(err);
+    if(req.user.accountType === 'Petani'){
+      try {
+        const farm = await Farm.findById(req.params.id);
+        res.json(farm);
+      } catch (error) {
+        const err = new Error("Farm not found.");
+        err.status = 404;
+        next(err);
+      }
     }
+    else {
+      res.status(400).json({ error: "Token is not valid" });
+    }
+    
   })
   .post((req, res, next) => {
     const err = new Error("POST method is not allowed.");
@@ -59,32 +74,43 @@ farmRouter
     next(err);
   })
   .put(async (req, res, next) => {
-    try {
-      const data = await Farm.findOneAndUpdate(
-        { _id: req.params.id },
-        req.body
-      );
-
-      res.json({ message: "Farm updated sucessfully!", data: req.body });
-    } catch (error) {
-      const err = new Error(error);
-      next(err);
+    if(req.user.accountType === 'Petani'){
+      try {
+        const data = await Farm.findOneAndUpdate(
+          { _id: req.params.id },
+          req.body
+        );
+  
+        res.json({ message: "Farm updated sucessfully!", data: req.body });
+      } catch (error) {
+        const err = new Error(error);
+        next(err);
+      }
+    } else {
+      res.status(400).json({ error: "Token is not valid" });
     }
+    
   })
   .delete(async (req, res, next) => {
-    try {
-      await Farm.findByIdAndDelete(req.params.id);
-      res.json({ message: "Farm deleted successfully." });
-    } catch (error) {
-      const err = new Error(error);
-      next(err);
+    if(req.user.accountType === 'Petani'){
+      try {
+        await Farm.findByIdAndDelete(req.params.id);
+        res.json({ message: "Farm deleted successfully." });
+      } catch (error) {
+        const err = new Error(error);
+        next(err);
+      }
+    } else {
+      res.status(400).json({ error: "Token is not valid" });
     }
+   
   });
 
 farmRouter
   .route("/:id/condition")
   .get(async (req, res, next) => {
-    const farm = await Farm.findById(req.params.id);
+    if(req.user.accountType === 'Petani'){
+      const farm = await Farm.findById(req.params.id);
 
     Promise.all([
       fetch(
@@ -117,6 +143,10 @@ farmRouter
         const err = new Error(error);
         next(err);
       });
+    } else {
+      res.status(400).json({ error: "Token is not valid" });
+    }
+    
   })
   .post((req, res, next) => {
     const err = new Error("POST method is not allowed.");
